moving row logic to recoil
